requirements.txt
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyserial>=3.5
joblib>=1.3.0
python-dateutil>=2.8.0
pytz>=2023.3
